requirements.txt
fastapi
uvicorn
pydantic
yfinance
pandas
numpy
requests
beautifulsoup4
python-dateutil
openai
tiktoken
python-dotenv
feedparser
pydantic>=2
cachetools>=5.3
tenacity>=8.2
httpx>=0.27
langchain
langchain-openai
langchain-community
faiss-cpu
streamlit
plotly
